requirements.txt
Django==4.2.7
django-grappelli==3.0.6
Pillow==10.0.0
djangorestframework==3.14.0
markdown==3.5.1
django-filter==23.5
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.2.2
google-auth==2.22.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.0.0
whitenoise==6.5.0